Fix memory to use the newer fixture organization
