+++ llama_stack/providers/tests/conftest.py
@@ -129,4 +129,5 @@ def pytest_itemcollected(item):
 pytest_plugins = [
     "llama_stack.providers.tests.inference.fixtures",
     "llama_stack.providers.tests.safety.fixtures",
+    "llama_stack.providers.tests.memory.fixtures",
 ]
