--- llama_stack/providers/tests/safety/provider_config_example.yaml
@@ -1,19 +1,0 @@
-providers:
-  inference:
-  - provider_id: together
-    provider_type: remote::together
-    config: {}
-  - provider_id: tgi
-    provider_type: remote::tgi
-    config:
-      url: http://127.0.0.1:7002
-   - provider_id: meta-reference
-     provider_type: meta-reference
-     config:
-       model: Llama-Guard-3-1B
-  safety:
-    - provider_id: meta-reference
-      provider_type: meta-reference
-      config:
-        llama_guard_shield:
-          model: Llama-Guard-3-1B
